switched to tape; added lock file;
