+++ package.json
@@ -7,7 +7,7 @@
     "example": "example"
   },
   "scripts": {
-    "test": "jasmine JASMINE_CONFIG_PATH=test/jasmine.json",
+    "test": "node ./test/test.js",
     "debug": "$NODE_DEBUG_OPTION jasmine JASMINE_CONFIG_PATH=test/jasmine.json"
   },
   "repository": {
@@ -27,8 +27,8 @@
   "devDependencies": {
     "@turf/helpers": "^3.6.3",
     "@turf/point-grid": "^3.14.0",
-    "jasmine": "2.6.0",
     "load-json-file": "^2.0.0",
+    "tape": "4.9.1",
     "write-json-file": "^2.0.0"
   }
 }
