+++ test/data/in/example-case.json
@@ -1,0 +1,14 @@
+{
+  "matrix":  [
+    [18, 13, 10,  9, 10, 13, 18],
+    [13,  8,  5,  4,  5,  8, 13],
+    [10,  5,  2,  1,  2,  5, 10],
+    [ 9,  4,  1, 12,  1,  4,  9],
+    [10,  5,  2,  1,  2,  5, 10],
+    [13,  8,  5,  4,  5,  8, 13],
+    [18, 13, 10,  9, 10, 13, 18],
+    [18, 13, 10,  9, 10, 13, 18]
+  ],
+  "lowerBand": 4.5,
+  "upperBand": 9
+}
