+++ test/data/in/issue#6-0_3.json
