--- test/jasmine.json
@@ -1,11 +1,0 @@
-{
-  "spec_dir": "test/specs/",
-  "spec_files": [
-    "*.spec.js"
-  ],
-  "helpers": [
-    "beforeTests/**/*.js"
-  ],
-  "stopSpecOnExpectationFailure": false,
-  "random": false
-}
